Create README.md
Four data sets provided in Excel (US Census, CDC Influenza Patient Visits, CDC Influenza Deaths by State, and Flu Shots in Children). After cleaning and wrangling the data, it had to be determined how many medical workers should be sent to each state in order to help the state combat the flu. Based off the number of deaths/hospital visits, population of the state, and flu shots administered it had to be determined which regions needed the most assistance and also when the assistance was needed. A higher emphasis was placed on areas that had a large vulnerable to influenza population.